fix cursor on search

+++ src/lib.rs
@@ -3,5 +3,5 @@ pub mod row;
 pub mod content;
 
 const VERSION: &str = "0.0.1";
-const TAB_STOP: u8 = 4;
+const TAB_STOP: usize = 4;
 const QUIT_TIMES: u8 = 2;
